// types/shared/payment.ts
export enum CARD_BRAND {
  VISA = 'Visa',
  AMERICAN_EXPRESS = 'American Express',
  DINERS_CLUB = 'Diners Club',
  JCB = 'JCB',
  MASTERCARD = 'MasterCard',
  UNION_PAY = 'UnionPay',
  UNKNOWN = 'Unknown',
}

export enum PAYMENT_METHOD {
  AMAZON = 'amazon',
  PAYPAL = 'paypal',
  AFFIRM = 'affirm',
  RESOLVE = 'resolve',
  KLARNA = 'klarna',
  IDEAL = 'ideal',
  BANCONTACT = 'bancontact',
  GOOGLE = 'google',
  APPLE = 'apple',
  CARD = 'card',
  PAYSAFECARD = 'paysafecard',
  MANUAL = 'manual',
}